trace-cmd: Do not have local builds of libtraceevent or libtracefs use system headers
If building locally, do not use system includes. This is needed for
kernelshark v1 to build libtraceevent and libtracefs. If the system header
files are used, there may be a mismatch with the frozen version of
libtraceevent or libtracefs that kernelshark will use, and the system
headers, and the build will fail.

Link: https://lore.kernel.org/linux-trace-devel/20201216232145.613591947@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>