trace-cmd: Add LOCAL_LIBS variable to force building the local libraries
In the process of changing the API for libtracefs before it becomes live,
add LOCAL_LIBS make command line variable to allow overriding the check if
libtracevent and libtracefs exist. If LOCAL_LIBS=1 is set on the make
command line, then the local versions of the libraries are built.

This is mostly for debug purposes, so it does not need to be documented.

Link: https://lore.kernel.org/linux-trace-devel/20201221095845.1a6f2aa7@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>