kernel-shark: Use mouse wheel to scroll the graphs
The current zoom-in/zoom-out function of the mouse wheel will be
performed only when rolling the mouse wheel and simultaneously
pressing "Ctrl", while just rolling the wheel will move you up or
down in the scroll area of the graphs without zooming.

We want to introduce the change because zoom via Ctrl+ScrollWheel
is a binding familiar from the Web Browsers.

Link: https://lore.kernel.org/linux-trace-devel/20210125140539.1029915-2-y.karadz@gmail.com

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=211253
Suggested-by: FeRD (Frank Dana) <ferdnyc@gmail.com>
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>