trace-cmd: Remove APIs for pairing a tracing peer.
In host-guests tracing context, tracing peers are all machines which are
being traced at the same time, as part of a single tracing session. The
logic for matching and pairing tracing peers from the same session, when
opening trace files, is used to take decision whether to apply time
stamps correction on the tracing events, in order to be in the same time
space. As there is a change in that behaviour - now these corrections
are applied by default, there is no need for pairing logic. These APIs
are removed:
 tracecmd_pair_peer()
 tracecmd_unpair_peer()
Note that the peer information is still saved in the file metadata.

Link: https://lore.kernel.org/linux-trace-devel/20210304131112.204294-3-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>