trace-cmd report: Add last_timestamp array for instances
The logic to add a last_timestamp array to the reading of trace data for
multiple files failed to include the array for instances within a given
file. This caused the instance handles not to have a way to save the last
timestamp, and caused the output to quit due to reading a record on a CPU
past the last CPU in the last_timestamp array.

Link: https://lore.kernel.org/linux-trace-devel/20210305224750.351786306@goodmis.org

Fixes: db36a61d2 ("trace-cmd: Fix last_timestamp logic to handle multiple files")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>