trace-cmd: Move tracecmd_write_cmdlines() out of tracecmd_append_cpu_data()
When the saved_cmdlines was moved to after the recording, as it makes sense
because they are created during the recording, the saving was just tossed
into tracecmd_append_cpu_data() as that was called at the end of recording.

Unfortunately, the trace-cmd restore (as well as trace-cmd split) used the
tracecmd_append_cpu_data() but expecting it not store the saved_cmdlines.
This broke both of them.

Now that there's an API called tracecmd_write_cmdlines(), have those that
need it call it directly, and remove it out of tracecmd_append_cpu_data().

Note, although this can help the trace-cmd restore code, it appears that the
code for trace-cmd split may still be broken, and needs to be fixed.

Link: https://lore.kernel.org/linux-trace-devel/20210301143857.098048355@goodmis.org
Link: https://lore.kernel.org/linux-trace-devel/20210305225947.541189414@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>