trace-cmd: Move the output state updates into the functions that change the state
It makes more sense to have the functions that change the state of the
descriptor to change the value that stores the state. This makes it more
robust in case these functions are called by something other than
create_file_fd(). That way the state changes with the update, and this
removes the dependency on create_file_fd with the state changes.

Link: https://lore.kernel.org/linux-trace-devel/20210301143857.244340263@goodmis.org
Link: https://lore.kernel.org/linux-trace-devel/20210305225947.694840219@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>