trace-cmd: Have tracecmd_read_headers() specify the state to read up to
As trace-cmd restore needs to read only a portion of the headers, it needs a
way to not read all of them. Export the tracecmd_file_state enums and pass
that as another parameter to tracecmd_read_headers(), in which it will stop
at the state requested.

If zero is passed in as the state, tracecmd_read_headers() will act in its
old behavior and read everything it can.

Now trace-cmd restore can call tracecmd_read_headers() and have it stop
after it reads the cmdlines.

Link: https://lore.kernel.org/linux-trace-devel/CAPpZLN5fdY+rb6s=X6bNykiqrjuTprvZxay9iDSEQxJXHURh3A@mail.gmail.com/
Link: https://lore.kernel.org/linux-trace-devel/20210305225948.143320048@goodmis.org

Suggested-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>