trace-cmd: Update make-trace-cmd.sh to handle pkg-config without --with-path
Debian and Ubuntu's pkg-config does not support the "--with-path" option.
Check first if it is supported, and only use it when it is.

Set PKG_CONFIG_PATH to inform pkg-config where to search.

Note, this script may not be the best way to create packaging and there may
be better ways to do so.

Link: https://lore.kernel.org/linux-trace-devel/20210309122216.4ce353e4@gandalf.local.home

Reported-by: pierre.gondois@arm.com
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=212149
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>