trace-cmd: Move time sync logic in the trace-cmd library
The whole timestamp synchronization logic is moved in the trace-cmd
library. It is compiled regardless of if vsock is supported or not,
although it will not work if thers is no vsock support. The logic is
abstracted, the application receives a time sync context which
represents the synchronization session. These library APIs can be used
to control the synchronization session:
  tracecmd_tsync_init()
  tracecmd_tsync_proto_getall()
  tsync_proto_is_supported()
  tracecmd_tsync_with_host()
  tracecmd_tsync_with_host_stop()
  tracecmd_tsync_with_guest()
  tracecmd_tsync_with_guest_stop()
  tracecmd_tsync_free()
  tracecmd_tsync_get_offsets()
  tracecmd_tsync_get_session_params()
  tracecmd_write_guest_time_shift()

Link: https://lore.kernel.org/linux-trace-devel/20210317113936.271143-2-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>