trace-cmd: Use open() and not fopen() for set_plugin_instance()
After hitting an error in writing to the "current_tracer" file that was not
reported by trace-cmd, it was due to the use of fopen(). Unless otherwise
specified, fwrite() will buffer writes, which can hide errors when writing
to the file.

Instead use open() and write() where the return gives an immediate result of
success or failure. There's no real reason to use fopen() anyway, as the
writes are rather trivial.

Link: https://lore.kernel.org/linux-trace-devel/20210325151000.76f836e1@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>