trace-cmd: Do not free tsync from thread
The freeing of an object should happen in the same context as where it is
created. The tsync object is created in the main thread and passed to the
guest and host handling threads. It should be closed and freed by the thread
that created it, especially since that thread accesses the tsync object
after creating the thread. Having the created thread free it would cause all
sorts of dangerous race conditions.

Link: https://lore.kernel.org/linux-trace-devel/20210331230728.350421194@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>