trace-cmd record: Have -d only touch the instance its for
The -d option for trace-cmd record disables the function tracing option.
It currently only does it for the top level, which does not affect the
instances. It should only disable the instances it is for and not touch
the top level if the top level is not part of the command line.

Also have set_ftrace return the result of set_ftrace_proc() if it was
called.

Link: https://lore.kernel.org/linux-trace-devel/20210407175427.536f0423@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>