trace-cmd libtracecmd: Fix the man page building and installing
When trying to do a make install_doc, the make hung due to the hacky way
of searching for the docbook.xsl on the system. As libtracecmd (and
libtracefs and libtraceevent) have a better way of creating and installing
the documentation, update the trace-cmd documentation to do the same. In
the mean time, also consolidate some of the build of the libtracecmd with
trace-cmd man pages.

Also fixed some leftover references to libtracefs.

Link: https://lore.kernel.org/linux-trace-devel/20210409233750.3778b5d4@oasis.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>