trace-cmd report: Update tep_func_repeat_format to match ts
With the new func_repeats event coming to the kernel, libtraceevent is
adding a way to process that event, and will display the timestamp of the
last function that repeated.

To make sure the displaying of that format matches the displaying of the
timestamps generated, update the tep_func_repeat_format variable to keep it
the same.

As the tep_func_repeat_format is a weak variable in the function plugin of
libtraceevent, it only needs to be defined to update it. Defining the
variable will not limit trace-cmd from using older libtraceevent libraries
that do not yet have this support.

Link: https://lore.kernel.org/linux-trace-devel/20210413173000.3e377bcc@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>