trace-cmd stat: Add -o option to show options
Add "-o" to trace-cmd stat to list the options that are enabled or disable
(like it does with trace-cmd list). The difference with stat is that it
allows you to see the available options for an instance, where as
trace-cmd list does not.

 trace-cmd stat -B foo -o

Will list the options that are set for instance foo.

Link: https://lore.kernel.org/linux-trace-devel/20210416172331.3870833-4-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>