trace-cmd library: Add tracecmd_info() log function
The pr_stat() function is removed from the traceevent library. This function
was used by trace-cmd to print statistics. Introduced tracecmd_info() log
function and replaced pr_stat() with it in the trace-cmd library.
However, pr_stat() is used also in the trace-cmd application. In order no
to change the existing behaviour, added a static pr_stat() implementation for
that usage.

Link: https://lore.kernel.org/linux-trace-devel/20210507095333.1080910-3-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
[ renamed local pr_stat() to print_event() to avoid name space conflicts ]
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>