trace-cmd: Auto discover running KVM VMs using virsh
The "trace-cmd record -A <guest>" option, used for guest tracing,
accepts VM name as parameter. But trace-cmd needs cid of that VM, in
order to perform the tracing. The old logic searches for a process with
name "qemu-system-" and parses its command line, to get guest name to
cid relation. Using a hardcoded process name and relying on specific
process command line arguments is not reliable way.
There is no API to get VM name to cid relation, but virsh can be used
instead of the old logic. As virsh is the official user interface to
libvirt, it should be more reliable than the old approach.

Link: https://lore.kernel.org/linux-trace-devel/20210514041141.11319-1-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>