trace-cmd: Do not enforce upper limit of verbose id
Applications that have a "verbose" option, do not usually set an upper
limit of what can be passed in. If the max is -V6, trace-cmd should not
error if someone passes in -V7. The way this is handled, is just set the
verbosity to the highest level if the number supplied is above the highest
level.

Link: https://lore.kernel.org/linux-trace-devel/20210625201511.1206465-4-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>