trace-cmd split: Assert if the calculated record size is too big
With a discovery of a bug that caused a record to be written passed the
end of a page (and this was a possible memory corruption bug), check that
the calculated length is no bigger than the record it is copying. If it
is, then crash, as this can cause the data to write pass the allocated
page.

Now that bug would crash on the split command with:

  Bad calculation of record len (expect:116 actual:120)

Link: https://lore.kernel.org/linux-trace-devel/20210628222609.01ea12ad@oasis.local.home/
Link: https://lore.kernel.org/linux-trace-devel/20210628224042.0ecc2030@oasis.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>