trace-cmd: Do not use local headers when system libraries are present
As the external libraries are starting to become more enhanced, they are
no longer the same as the local headers and will break the build if the
two exists. The problem is, there's locations in the C code that uses
"traceevent/event-parse.h" instead of just "event-parse.h", which picks up
the local headers even if the libraries are installed. Remove the extra
path name from the includes to not pick them up locally when the system
libraries exist.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>