trace-cmd: For libtraceevent 1.3 have pr_info equal tep_info
In version libtracevent 1.3, pr_stat() turned to tep_info(), because
"pr_stat" was too generic. Although a compiled version of trace-cmd may
still work because of the use of weak functions, building this version of
trace-cmd may cause several warnings because warning is not declared.

Since pr_stat is called out as deprecated, simply doing a:

  -Dpr_stat=tep_info

is not sufficient, as the header that deprecates pr_stat will also be
affected, and you still get the warning, that is confusing.

  "tep_info" is deprecated, use tep_info instead.

Instead, change all callers of pr_stat() to pr_info() and have the
Makefile do -Dpr_info=pr_stat when libtraceevent is less than 1.3, and
-Dpr_info=tep_info when it is equal or greater to 1.3.

Link: https://lore.kernel.org/linux-trace-devel/20210716225712.1bc5aa77@rorschach.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>