trace-cmd: Remove traceevent/ and tracefs/ from include headers
trace-cmd now depends on the external libraries of both libtraceevent and
libtracefs and uses pkg-config to find their headers. As pkg-config will
find the full path to the headers, do not include parts of the path when
including them, because they may change, and the hardcoded path in the
code will not.

Link: https://lore.kernel.org/linux-trace-devel/20210726162857.487153-2-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>