trace-cmd: Fix out of range comparison
Comparing an unsigned int with a uint64_t causes zero extension of the
unsigned int. This doesn't allow the unsigned int to ever equal LONG_MIN
or LONG_MAX, i.e the branch is always false.

This issue was caught by clang:

tracecmd/trace-vm.c:340:37: error: result of comparison of constant 9223372036854775807 with expression of type 'unsigned int' is always false [-Werror,-Wtautological-constant-out-of-range-compare]
                if ((cid_id == LONG_MIN || cid_id == LONG_MAX) && errno == ERANGE)
                                           ~~~~~~ ^  ~~~~~~~~
tracecmd/trace-vm.c:340:15: error: result of comparison of constant -9223372036854775808 with expression of type 'unsigned int' is always false [-Werror,-Wtautological-constant-out-of-range-compare]
                if ((cid_id == LONG_MIN || cid_id == LONG_MAX) && errno == ERANGE)

Link: https://lore.kernel.org/linux-trace-devel/20210805043610.2116406-1-irogers@google.com

Signed-off-by: Ian Rogers <irogers@google.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>